+++ pyphyschemtools/__init__.py
@@ -3,8 +3,8 @@
 The pyphyschemtools library provides a comprehensive suite of utilities for Physical Chemistry, ranging from spectroscopic unit management to kinetic modeling and cheminformatics.
 """
 
-__version__ = "0.8.1"
-__last_update__ = "2026-02-27"
+__version__ = "0.8.2"
+__last_update__ = "2026-03-04"
 
 import importlib
 import importlib.util